feat: add .gitignore entry for ws-settings.zip and docker/.secret/
Add entries for `ws-settings.zip` and `docker/.secret/` to `.gitignore` to prevent them from being tracked by Git.

The `ws-settings.zip` file is generated by the application and contains user-specific settings, which should not be version-controlled.

The `docker/.secret/` directory contains sensitive information, such as passwords and API keys, that should not be included in the Git repository.

This commit improves the overall security of the repository by preventing these files from being tracked and potentially exposed.