feat: Add support for custom data file directory and name
The `.env` file now includes `DATA_FILE_DIR` and `DATA_FILE_NAME`
variables for specifying the path and name of the data file to be used.
In the Dockerfile, the `DATA_FILE_DIR` argument is passed to the
environment variable `DATA_FILE_DIR` for use in the container. The
`entrypoint.sh` script now checks if the node modules are installed and
installs them if not. The `docker-compose.yml` file has been updated to
use the `DATA_FILE_DIR` and `DATA_FILE_NAME` variables in the volume
mounts. Additionally, the `LOCAL_KEYS_FOLDER` variable has been removed
from the `.env` file and the volume mount in the `docker-compose.yml`
file, and replaced with a `ssl_keystore` volume for mounting SSL
certificates.