feat: Implement SQLite backend for KeyStorage
This commit adds support for SQLite backend to KeyStorage class. A new `createDatabase` method is added to `SqliteStorage` class to initialize the database and it is called inside the `getInstance` method. The `constructor` of the `SqliteStorage` class is also modified to take a `db` parameter which is used to create a new instance of the class. Additionally, the console log message for initializing the backend is updated to be more descriptive.