feat: add custom query method to StorageStrategy interface
This commit adds a new abstract method called `custom` to the
`StorageStrategy` interface. This method accepts a string query and
returns a promise with an array of objects with string and number
properties. The commit also includes changes to the `SQLite` backend
to validate the data returned by the `select` method and use a
temporary in-memory database for testing purposes.