feat: Add SSL support to SoakpServer
Add SSL support to SoakpServer by adding the `iniSSL` method to initialize
SSL credentials using the private key and certificate files. Use `https`
module to create an SSL server. Call `iniSSL` method in `start` method to
initialize SSL and listen to both HTTP and HTTPS requests. Update the
constructor of SoakpServer to remove the `configuration` parameter and
add `cors` middleware to the application. Remove the `init` method and
pass the `keyStorage` parameter to `start` method.