Merge develop to master (#7)
* chore: Remove exposed keys

* chore: add LICENSE.md file with Lehcode.com copyright and MIT License

* chore: update Dockerfile and docker-compose.yml configurations

- Update Dockerfile to include chown for both the workdir and data_file_dir
- Remove VOLUME from myca Dockerfile
- Update entrypoint.sh to include chown for data_file_dir and remove comments
- Update docker-compose.yml to change OPENAI_API_ORG to OPENAI_API_ORG_ID and update volumes to reflect changes in SSL_CERTS_DIR and DATA_FILE_DIR

* feat: Add OpenAI API request parameters

Added `temperature` and `maxTokens` parameters to the OpenAI request interface for more flexible AI completion. Updated the `SoakpProxy` class to use these new parameters in the `request` method instead of hardcoded values. Also updated the `handleJwtRefresh` method in `SoakpServer` to use environment variables for `apiOrgKey` and `apiKey`. Finally, changed the `/openai/query` endpoint to `/openai/completion` and added a new endpoint `/openai/models` for fetching the available models from the OpenAI API.

* feat(env): update OpenAI API credentials in .env.dist file

* chore(README): Update package description and add documentation link

Update the package description to reflect that SOAKP is both a Node.js application and library. Also, add a link to the full documentation for the package. This commit does not include any code changes.