build: update Dockerfile and entrypoint.sh
Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.