Develop (#17)
* feat: Add SSL certificate configuration, deployment options, and update documentation

This commit adds information on SSL certificate configuration in the README file and updates the Certbot.md file with a corrected command to start the Certbot service. The index.md file has been updated with a new section on features, and the name of a file has been changed from UsageScenarios.md to Scenarios.md. Additionally, a new file has been added under the deployment folder, CompileJS.md, which provides instructions for compiling JavaScript files.

Added instructions for configuring SSL certificates, including generating self-signed certificates for local development and using existing certificates for localhost. Also added documentation for deployment options, including hosting on cloud platforms or local machines, generating JavaScript files for deployment, and deploying to AWS Lambda for serverless deployment.

* Fix merge conflict

* fix: Generate SSL certificates using myca Docker image

This commit generates SSL certificates using the myca Docker image.
The Dockerfile in `docker/myca` has been updated to include a new build
argument `host_docker_gid`, which specifies the GID of the "docker"
group of the host machine. The entrypoint script has been modified to
copy the generated certificates to the directory specified in the
`SSL_CERT_DIR` environment variable. The `docker-compose.yml` file has
also been updated to use the new `SSL_CERT_DIR` variable instead of
`SSL_CERTS_DIR`.

* build: update Dockerfile and entrypoint.sh for host user customization

This commit updates the Dockerfile to allow customization of the host
user's UID and username, in addition to their password. It also updates
the entrypoint.sh script to log the contents of the SSL cert directory
in debug mode.

Specifically, this commit:
- Adds two new build arguments, `host_user_uid` and `host_user_name`,
  for the host user's UID and username, respectively
- Replaces the `useradd` command in the Dockerfile with a command that
  uses the new build arguments to create the host user
- Updates the `sudoers` file in the Dockerfile to allow the host user to
  run certain commands without a password
- Replaces references to the `node` user in the Dockerfile and
  entrypoint.sh with references to the new host user
- Adds a conditional in the entrypoint.sh script to log the contents of

* feat: Update OpenAIRequestInterface and SoakpServer

- Update OpenAIRequestInterface to include new properties for messages, engineId,
  max_tokens, and suffix
- In SoakpServer, update the use of OpenAIRequestInterface properties to reflect
  the changes made in the interface definition
- Rename iniSSL to initSSL in SoakpServer
- Fix KeyStorage fetchJWT method to use findOne instead of find

* feat(env): update OPENAI_API_ORG_ID and SERVER_HOST in .env.dist file

* build: update Dockerfile and entrypoint.sh

Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.