Merge develop (#22)
* build: update Dockerfile and entrypoint.sh for host user customization

This commit updates the Dockerfile to allow customization of the host
user's UID and username, in addition to their password. It also updates
the entrypoint.sh script to log the contents of the SSL cert directory
in debug mode.

Specifically, this commit:
- Adds two new build arguments, `host_user_uid` and `host_user_name`,
  for the host user's UID and username, respectively
- Replaces the `useradd` command in the Dockerfile with a command that
  uses the new build arguments to create the host user
- Updates the `sudoers` file in the Dockerfile to allow the host user to
  run certain commands without a password
- Replaces references to the `node` user in the Dockerfile and
  entrypoint.sh with references to the new host user
- Adds a conditional in the entrypoint.sh script to log the contents of

* feat(env): update OPENAI_API_ORG_ID and SERVER_HOST in .env.dist file

* build: update Dockerfile and entrypoint.sh

Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.

* chore(package.json): update keywords and add bugs url

Update the keywords in package.json to include "soakp", "openai", "key",
"browser", and "gpt". Also add a "bugs" object with a URL to the issue
tracker on GitHub.

* feat(package.json): add line breaks to keywords array

Add line breaks to the keywords array for better readability. The keywords
are "soakp", "openai", "key", "browser", and "gpt".

* Fix merge conflict

* fix: Generate SSL certificates using myca Docker image

This commit generates SSL certificates using the myca Docker image.
The Dockerfile in `docker/myca` has been updated to include a new build
argument `host_docker_gid`, which specifies the GID of the "docker"
group of the host machine. The entrypoint script has been modified to
copy the generated certificates to the directory specified in the
`SSL_CERT_DIR` environment variable. The `docker-compose.yml` file has
also been updated to use the new `SSL_CERT_DIR` variable instead of
`SSL_CERTS_DIR`.

* build: update Dockerfile and entrypoint.sh for host user customization

This commit updates the Dockerfile to allow customization of the host
user's UID and username, in addition to their password. It also updates
the entrypoint.sh script to log the contents of the SSL cert directory
in debug mode.

Specifically, this commit:
- Adds two new build arguments, `host_user_uid` and `host_user_name`,
  for the host user's UID and username, respectively
- Replaces the `useradd` command in the Dockerfile with a command that
  uses the new build arguments to create the host user
- Updates the `sudoers` file in the Dockerfile to allow the host user to
  run certain commands without a password
- Replaces references to the `node` user in the Dockerfile and
  entrypoint.sh with references to the new host user
- Adds a conditional in the entrypoint.sh script to log the contents of

* build: update Dockerfile and entrypoint.sh

Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.

* chore(package.json): update keywords and add bugs url

Update the keywords in package.json to include "soakp", "openai", "key",
"browser", and "gpt". Also add a "bugs" object with a URL to the issue
tracker on GitHub.

* feat(package.json): add line breaks to keywords array

Add line breaks to the keywords array for better readability. The keywords
are "soakp", "openai", "key", "browser", and "gpt".

* feat: add @types/cors to devDependencies

This commit adds "@types/cors" to the devDependencies in package.json.