Merge develop (#26)
* build: update Dockerfile and entrypoint.sh for host user customization

This commit updates the Dockerfile to allow customization of the host
user's UID and username, in addition to their password. It also updates
the entrypoint.sh script to log the contents of the SSL cert directory
in debug mode.

Specifically, this commit:
- Adds two new build arguments, `host_user_uid` and `host_user_name`,
  for the host user's UID and username, respectively
- Replaces the `useradd` command in the Dockerfile with a command that
  uses the new build arguments to create the host user
- Updates the `sudoers` file in the Dockerfile to allow the host user to
  run certain commands without a password
- Replaces references to the `node` user in the Dockerfile and
  entrypoint.sh with references to the new host user
- Adds a conditional in the entrypoint.sh script to log the contents of

* feat(env): update OPENAI_API_ORG_ID and SERVER_HOST in .env.dist file

* build: update Dockerfile and entrypoint.sh

Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.

* chore(package.json): update keywords and add bugs url

Update the keywords in package.json to include "soakp", "openai", "key",
"browser", and "gpt". Also add a "bugs" object with a URL to the issue
tracker on GitHub.

* feat(package.json): add line breaks to keywords array

Add line breaks to the keywords array for better readability. The keywords
are "soakp", "openai", "key", "browser", and "gpt".

* Fix merge conflict

* fix: Generate SSL certificates using myca Docker image

This commit generates SSL certificates using the myca Docker image.
The Dockerfile in `docker/myca` has been updated to include a new build
argument `host_docker_gid`, which specifies the GID of the "docker"
group of the host machine. The entrypoint script has been modified to
copy the generated certificates to the directory specified in the
`SSL_CERT_DIR` environment variable. The `docker-compose.yml` file has
also been updated to use the new `SSL_CERT_DIR` variable instead of
`SSL_CERTS_DIR`.

* build: update Dockerfile and entrypoint.sh for host user customization

This commit updates the Dockerfile to allow customization of the host
user's UID and username, in addition to their password. It also updates
the entrypoint.sh script to log the contents of the SSL cert directory
in debug mode.

Specifically, this commit:
- Adds two new build arguments, `host_user_uid` and `host_user_name`,
  for the host user's UID and username, respectively
- Replaces the `useradd` command in the Dockerfile with a command that
  uses the new build arguments to create the host user
- Updates the `sudoers` file in the Dockerfile to allow the host user to
  run certain commands without a password
- Replaces references to the `node` user in the Dockerfile and
  entrypoint.sh with references to the new host user
- Adds a conditional in the entrypoint.sh script to log the contents of

* build: update Dockerfile and entrypoint.sh

Update the Dockerfile to add new arguments and environment variables.
Add a new group for Docker and another group for the host user.
Create the home directory for the host user and add it to the Docker group.
Update the entrypoint.sh script to use the new host user UID and GID.

* chore(package.json): update keywords and add bugs url

Update the keywords in package.json to include "soakp", "openai", "key",
"browser", and "gpt". Also add a "bugs" object with a URL to the issue
tracker on GitHub.

* feat(package.json): add line breaks to keywords array

Add line breaks to the keywords array for better readability. The keywords
are "soakp", "openai", "key", "browser", and "gpt".

* feat: add @types/cors to devDependencies

This commit adds "@types/cors" to the devDependencies in package.json.

* refactor: Fix typings

* feat: update .eslintrc.json rules

Update the indent rule to warn instead of error, and adjust the options
for FunctionDeclaration and FunctionExpression. Also ignore chains with a
depth of 2 for newline-per-chained-call.

* feat:  Added ./dict folder

* feat: add .npmrc to .gitignore

* feat(tsconfig): update compiler options for Node16 and es2018 target

* 1.0.2

* build(tsconfig): add Jest config, declaration and declarationMap options

* chore:  update dist

* feat: add null check to backend in KeyStorage.ts

Add null check to backend in KeyStorage.ts to prevent null reference
errors. This is achieved by adding a null check to the backend
property in the constructor and in methods that use it.

* feat: Update tsconfig.json to include source map configuration

Changed the "inlineSourceMap" property to "false" and added a new property
"sourceMap" and set it to "true" in tsconfig.json file. This will generate a
separate source map file during compilation.

* feat: Implement fallback options and storage configuration

Implemented fallback options for server configuration and storage configuration for KeyStorage. The `fallback` object is now used as a default value for missing or undefined keys in configuration. The `StorageConfigInterface` now includes `tableName`, `dbName`, `dataFileDir`, and `lifetime` properties.

Refactored `DbSchemaInterface` to extend `Object` for better typing. Updated `KeyStorage` constructor and `getInstance` method to accept the new `StorageConfigInterface` object and removed `StorageType`.

* build: update .gitignore to exclude unnecessary directories and files

* build: update tsconfig, package.json, and .babelrc

- Update tsconfig.json to use "node" module instead of "commonjs"
- Remove "dom" from lib in tsconfig.json
- Update include in tsconfig.json to use "**/*.ts"
- Add "babel-plugin-transform-remove-console" to dependencies in package.json
- Update .babelrc to remove "node" target from presets
- Add "transform-remove-console" plugin to production environment in .babelrc
- Enable comments cleanup for production mode JS in .babelrc.

* feat: refactor import statements and fix method naming

Refactored import statements in index.ts and SoakpProxy.ts to use named imports instead of default imports. Also fixed method naming in SoakpProxy.ts to use makeRequest instead of request.

* style(.prettierrc): remove unnecessary "semi" option in TypeScript and JSON parsers

* chore: update distribution build

* feat: update author and copyright information

Update author and copyright information in all relevant files to reflect the correct information.

* feat: update dotenv dependency to 16.1.4 in package.json file

* build: update tsconfig.json

* feat: add author and copyright information to files

Add author and copyright information to relevant files in the project.

* feat: update build

* feat: add yarn-error.log to .gitignore file

The commit adds yarn-error.log to the .gitignore file to prevent it from being
tracked in future commits.