Merge gh-pages (#39)
* feat(certbot): Update port variable name and clarify SSL certificate validity

Changed the name of SERVER_PORT to SECURE_PORT in .env file to better reflect its usage.
Also updated the documentation to clarify that the SSL certificate generated by Certbot is completely valid and can be used in production environments.

* feat: update README.md

* feat: add minima theme to _config.yml file

* feat: add updated information and features to README and docs

The commit features updated information to the README and documentation files.

- Update the `soakp-client` link in the README to the correct NPM package
- Refactor and rephrase some sentences in the README for clarity
- Update the `soakp-client` link in the docs to the correct GitHub repository
- Add details on SSL certificates generated by Certbot in the docs
- Reorder the features in the docs to match the README
- Update the formatting in both files

No breaking changes or new features are added.

* feat(docs): rename _config.yml to docs/_config.yml

* feat(docs): rename _config.yml to docs/_config.yml

* feat(docs): rename _config.yml to docs/_config.yml

* feat(docs): rename _config.yml to docs/_config.yml

* feat(docs): rename _config.yml to docs/_config.yml

* feat: change default branch to gh-pages in jekyll-gh-pages workflow

Changed the default branch to 'gh-pages' in the jekyll-gh-pages
workflow configuration file. This ensures that the workflow runs on
pushes targeting the 'gh-pages' branch instead of the 'master' branch.
This commit also allows manual triggering of the workflow from the
Actions tab.