Add NPM and GitHub Packages publish action (#47)
* feat: add remote deployment to GitHub Packages and NPM

Added a new workflow to deploy releases to GitHub Packages and NPM. The workflow runs tests using node on release creation and publishes the package to both GitHub Packages and NPM. The workflow also includes linting and testing steps. The NODE_AUTH_TOKEN environment variable is used to authenticate the deployment to both GitHub Packages and NPM.

* feat: add push trigger for publish branch in npm-publish.yml file

* build: enable publishing to NPM and GitHub Packages

This commit updates the `.github/workflows/npm-publish.yml` file to remove the
`push` trigger for the `publish` branch and adds a `release` trigger for
`published` types. It also removes the `--dry-run` flag from the `npm publish`
command, allowing for actual publishing to occur.

Additionally, a new workflow is added in `.github/workflows/npm-publish-dry-run.yml`
which performs a dry run of the publishing process on the `develop` branch. This
workflow runs `npm lint` and `npm test` in addition to the `npm publish --dry-run`
command, and is intended for use as a pre-publish check.

Both workflows now use Node.js version 16 and have been tested on Ubuntu.