Add NPM and GitHub Packages publish action (#47) (#48)
* feat: add NPM and GitHub Packages publish action

* feat: add remote deployment to GitHub Packages and NPM

Added a new workflow to deploy releases to GitHub Packages and NPM. The workflow runs tests using node on release creation and publishes the package to both GitHub Packages and NPM. The workflow also includes linting and testing steps. The NODE_AUTH_TOKEN environment variable is used to authenticate the deployment to both GitHub Packages and NPM.

* feat: add push trigger for publish branch in npm-publish.yml file

* build: enable publishing to NPM and GitHub Packages

This commit updates the `.github/workflows/npm-publish.yml` file to remove the
`push` trigger for the `publish` branch and adds a `release` trigger for
`published` types. It also removes the `--dry-run` flag from the `npm publish`
command, allowing for actual publishing to occur.

Additionally, a new workflow is added in `.github/workflows/npm-publish-dry-run.yml`
which performs a dry run of the publishing process on the `develop` branch. This
workflow runs `npm lint` and `npm test` in addition to the `npm publish --dry-run`
command, and is intended for use as a pre-publish check.

Both workflows now use Node.js version 16 and have been tested on Ubuntu.

* build: update npm commands to use yarn

This commit updates the npm commands in the .github/workflows/npm-publish-dry-run.yml file to use yarn instead. Specifically, `npm ci` is changed to `yarn install --frozen-lockfile`, `npm lint` is changed to `yarn lint`, and `npm test` is changed to `yarn test`. This should improve consistency and reliability in our development process.

* build(package.json): update dependencies

Updated the "test" script to run Jest in continuous integration mode with
the "test-ci" script. Also, optimized dependencies.

* feat: update npm-publish-dry-run.yml

This commit updates the npm-publish-dry-run.yml file by changing the test command from "yarn test" to "yarn test-ci".

* feat: update jest config

The jest.config.js file has been modified to ignore .js files in the testPathIgnorePatterns array.

* style: update .prettierrc settings for JSON files

Remove "tabWidth" option and set "singleQuote" to false for consistency.

* build: upgrade npm commands to yarn in npm-publish.yml file