docs: add jsdoc to static method to create SQLite database
Adds description to a static method `createDatabase` to the `SqliteStorage` class that
creates a SQLite database. The method takes `dbName`, `tableName`, and
`dbFile` as parameters and returns a promise that resolves with a
`Database` object. If `SQLITE_MEMORY_DB` environment variable is set to
`yes`, an in-memory database is created. If `SQLITE_RESET` environment
variable is set to `yes`, the existing database is deleted before
creating a new one. If an error occurs, it is logged to the console.