feat: Add docker-compose.testing.yml
This commit adds the docker-compose.testing.yml file, which contains the configuration for testing the application. It includes the definition of services such as the server and myca, along with their respective environment variables and dependencies. The myca service uses the lehcode/ssl-myca:latest image and has additional build arguments for debugging and generating SSL certificates. The volumes and networks configurations are also included.