feature/finetunes (#151)
* feat: Add fine-tunes API endpoints

* feat(SoakpProxy): Add concatTxt2Jsonlines method

This commit adds the `concatTxt2Jsonlines` method to the `SoakpProxy`
class. This method is used to concatenate an array of text files into a
single JSONL file suitable for OpenAI model fine-tuning. The method
takes in the `txtFiles` array of text files and the `completions` array
of completion strings. It also accepts an optional `concatBaseName`
parameter to specify the base name of the concatenated JSONL file. If
not provided, a default name will be generated.

Issue: feature/finetunes

* feat: Update Node.js version in .env.dist and .env.test files

Update the Node.js version in the .env.dist and .env.test files from 16.20.0 to 18.17.0.

Issue: feature/finetunes